requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
coverage==7.4.3
Deprecated==1.2.14
exceptiongroup==1.2.0
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.1.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.9.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
pluggy==1.4.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==4.1.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
wrapt==1.16.0
